+++ app/Http/Controllers/Admin/TipoController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers\Admin;
 
 use App\Http\Controllers\Controller;
+use App\Imovel;
 use App\Tipo;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Session;
@@ -58,6 +59,29 @@ class TipoController extends Controller
 
     public function deletar($id)
     {
+
+        if(Imovel::where('tipo_id', '=', $id)->count()){
+
+            $msg = "Nao e' possivel deletar essa tipo esses imoveis (";
+
+            $imoveis = Imovel::where('tipo_id', '=', $id)->get();
+
+            foreach ($imoveis as $imovel ) {
+                $msg .= "id: ".$imovel->id." ";
+            }
+
+            $msg .= ") estao relacionados";
+
+            Session::flash('mensagem', [
+                'msg' => $msg,
+                'class' => 'red white-text'
+            ]);
+
+            return redirect()->route('admin.tipos');
+
+        }
+
+
         Tipo::find($id)->delete();
 
         Session::flash('mensagem', [
